refactor: [#1268] move announce logic from udp server to udp_tracker_core package
