+++ src/packages/udp_tracker_core/mod.rs
@@ -1,2 +1,3 @@
+pub mod peer_builder;
 pub mod services;
 pub mod statistics;
